+++ R/get_predictions_polr.R
@@ -25,7 +25,7 @@ get_predictions_polr <- function(model, fitfram, ci.lvl, linv, typical, terms, m
   # case. just return predictions
   if (nrow(prdat) > nrow(fitfram) && ncol(prdat) == 1) {
     colnames(prdat)[1] <- "predicted"
-    return(rownames_as_column(prdat, var = "response.level"))
+    return(.rownames_as_column(prdat, var = "response.level"))
   }
 
   # bind predictions to model frame