+++ NEWS.md
@@ -6,11 +6,10 @@
 
 ## New supported models
 
-* `bracl`, `brmultinom` (package **brglm2**)
+* `bracl`, `brmultinom` (package **brglm2**) and models from packages **bamlss** and **R2BayesX**.
 
 ## General
 
-* `ggpredict()` now supports models from packages **bamlss** and **R2BayesX**.
 * Updated package dependencies.
 * `plot()` now uses dodge-position for raw data for categorical x-axis, to align raw data points with points and error bars geoms from predictions.